Actually create lighting effects.
Figured out the interface things needed for this and got it working elsewhere. This required an interface change, thus the version bump. Took advantage and tweaked the analog sensor for rationality.

Updated to latest Kotlin, fixed some dumb co-routine context stuff.

+++ src/main/kotlin/crackers/kobots/parts/movement/async/easing.kt
@@ -44,7 +44,7 @@ suspend fun easeTo(
     easingFn: EasingFunction = linear,
     steps: Int = 50,
 ) {
-    val stepDuration = (duration.inWholeMilliseconds / steps).coerceAtLeast(10).toLong()
+    val stepDuration = (duration.inWholeMilliseconds / steps).coerceAtLeast(10)
     for (i in 0..steps) {
         val t = i.toFloat() / steps
         val easedT = easingFn(t)
